Add files via upload
I'm uploading the first iteration of my portfolio site, I really like the idea of the black/white theme although it makes everything look a little bland. The text in the center is very much subject to change. One of my favorite things in CSS are transforms and transitions so I've tried to showcase them here with the links to GitHub and LinkedIn. I also tried to make the <h1> "appear" on the page but that didn't work out so it's now just a very light color which turns to black when hovered.  

Next up:
I'm going to add an About page. Now since I've not yet learned JavaScript I'll do it by creating another html file but as soon as I learn how to navigate multiple parts of one page (I'm not really sure how to call it) using JS it will be replaced by that