Added a "Contact me" section
Also changed the background color so the site doesn't look so basic. Made some improvements all-around. 
Not really sure if I like how the contact section looks. Initially I left the input boxes white but again, it looked too basic. Now they are a color similar to the background which has the benefit of making it look better (to me) but the downside of making it kind of blend in with the background