Added new "Project"
It's supposed to be a car selling/trim selecting type of thing but I'll see where it goes. Still very limited due to lack of JS so next step will probably be learning some of that.
Also discovered that "About" section looks different on a small screen (duh..) so I'll have to figure that one out.